Add unit tests for weekly planner retrieval, assignment, and duplicate prevention
